requirements.txt
pandas>=1.5
numpy
scikit-learn
matplotlib
requests
fpdf>=1.7.2
python-dateutil
pytz
joblib
streamlit